requirements.txt
python-telegram-bot>=22.7
h2>=4.0
edge-tts
langdetect
